fix embedded build
