+++ mavlink-bindgen/Cargo.toml
@@ -12,9 +12,14 @@ quote = "1"
 proc-macro2 = "1.0.43"
 lazy_static = "1.2.0"
 serde = { version = "1.0.115", optional = true, features = ["derive"] }
-clap = { version = "~4.3.24", default-features = false, features = ["std", "help", "usage", "error-context", "derive"] }
+clap = { version = "~4.3.24", optional = true, features = ["derive"] }
 thiserror = "1.0.56"
 
 # Required to support MSRV 1.65.0
-clap_lex = { version = "=0.5.0" }
-anstyle = { version = "=1.0.2" }
+clap_lex = { version = "=0.5.0", optional=true }
+clap_builder = { version = "~4.3.24", optional=true}
+anstyle = { version = "=1.0.2", optional=true }
+
+[features]
+cli = ["dep:clap", "dep:clap_lex", "dep:clap_builder", "dep:anstyle"]
+default = ["cli"]
